solution

+++ Database Programmability and  Transactions/Exercises/Future Value Function.sql
@@ -1,4 +1,4 @@
-CREATE FUNCTION ufn_calculate_future_value(sum DOUBLE(10, 4), rate DOUBLE(10, 2), years INT)
+CREATE FUNCTION ufn_calculate_future_value(sum DOUBLE(10, 4), rate DOUBLE(10, 4), years INT)
     RETURNS DECIMAL (10, 4)
     DETERMINISTIC
 BEGIN
